refactor: 파일 위치 변경
- Controller : domain -> controller
- LottoTest : domain밑으로 이동
