+++ src/main/java/lotto/controller/Controller.java
@@ -1,6 +1,7 @@
-package lotto.domain;
+package lotto.controller;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.domain.Money;
 import lotto.view.InputView;
 
 public class Controller {
@@ -24,6 +25,7 @@ public class Controller {
     }
 
     private void purchaseLotto(int money) {
+        int purchasableLotto = money / 1000;
 
     }
 }
